// packages/core/src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
	/** The `BigDecimal` scalar type represents signed fractional values with arbitrary precision. */
	BigDecimal: { input: any; output: any };
	/** ISO-8601 formatted date in UTC */
	DateTime: { input: any; output: any };
	/** The `Long` scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
	Long: { input: any; output: any };
};
// ...
export type CollectionInfo = {
	__typename?: 'CollectionInfo';
	/** Total items in the collection despite pagination. */
	totalItems?: Maybe<Scalars['Long']['output']>;
};
// ...
export type PageInfo = {
	__typename?: 'PageInfo';
	/** When paginating forwards, the cursor to continue. */
	endCursor?: Maybe<Scalars['String']['output']>;
	/** When paginating forwards, are there more items? */
	hasNextPage: Scalars['Boolean']['output'];
	/** When paginating backwards, are there more items? */
	hasPreviousPage: Scalars['Boolean']['output'];
	/** When paginating backwards, the cursor to continue. */
	startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type PriceSearchFilterInput = {
	/** Maximum price of the product. */
	maxPrice?: InputMaybe<Scalars['Float']['input']>;
	/** Minimum price of the product. */
	minPrice?: InputMaybe<Scalars['Float']['input']>;
};
// ...
export type Product = Node & {
	__typename?: 'Product';
	/** Absolute URL path for adding a product to cart. */
	addToCartUrl: Scalars['String']['output'];
	/**
	 * Absolute URL path for adding a product to customer's wishlist.
	 * @deprecated Deprecated.
	 */
	addToWishlistUrl: Scalars['String']['output'];
	/**
	 * The availability state of the product.
	 * @deprecated Use status inside availabilityV2 instead.
	 */
	availability: Scalars['String']['output'];
	/**
	 * A few words telling the customer how long it will normally take to ship this product, such as 'Usually ships in 24 hours'.
	 * @deprecated Use description inside availabilityV2 instead.
	 */
	availabilityDescription: Scalars['String']['output'];
	/** The availability state of the product. */
	availabilityV2: ProductAvailability;
	/** Brand associated with the product. */
	brand?: Maybe<Brand>;
	/** List of categories associated with the product. */
	categories: CategoryConnection;
	/** Product condition */
	condition?: Maybe<ProductConditionType>;
	/**
	 * Product creation date
	 * @deprecated Alpha version. Do not use in production.
	 */
	createdAt: DateTimeExtended;
	/** Custom fields of the product. */
	customFields: CustomFieldConnection;
	/** Default image for a product. */
	defaultImage?: Maybe<Image>;
	/** Depth of the product. */
	depth?: Maybe<Measurement>;
	/** Description of the product. */
	description: Scalars['String']['output'];
	/** Id of the product. */
	entityId: Scalars['Int']['output'];
	/** Gift wrapping options available for the product. */
	giftWrappingOptions: GiftWrappingConnection;
	/** Global trade item number. */
	gtin?: Maybe<Scalars['String']['output']>;
	/** Height of the product. */
	height?: Maybe<Measurement>;
	/** The ID of an object */
	id: Scalars['ID']['output'];
	/** A list of the images for a product. */
	images: ImageConnection;
	/** Inventory information of the product. */
	inventory: ProductInventory;
	/** Maximum purchasable quantity for this product in a single order. */
	maxPurchaseQuantity?: Maybe<Scalars['Int']['output']>;
	/** Metafield data related to a product. */
	metafields: MetafieldConnection;
	/** Minimum purchasable quantity for this product in a single order. */
	minPurchaseQuantity?: Maybe<Scalars['Int']['output']>;
	/** Manufacturer part number. */
	mpn?: Maybe<Scalars['String']['output']>;
	/** Name of the product. */
	name: Scalars['String']['output'];
	/**
	 * Product options.
	 * @deprecated Use productOptions instead.
	 */
	options: OptionConnection;
	/** Relative URL path to product page. */
	path: Scalars['String']['output'];
	/** Description of the product in plain text. */
	plainTextDescription: Scalars['String']['output'];
	/**
	 * The minimum and maximum price of this product based on variant pricing and/or modifier price rules.
	 * @deprecated Use priceRanges inside prices node instead.
	 */
	priceRanges?: Maybe<PriceRanges>;
	/** Prices object determined by supplied product ID, variant ID, and selected option IDs. */
	prices?: Maybe<Prices>;
	/** Product options. */
	productOptions: ProductOptionConnection;
	/** Related products for this product. */
	relatedProducts: RelatedProductsConnection;
	/** Summary of the product reviews, includes the total number of reviews submitted and summation of the ratings on the reviews (ratings range from 0-5 per review). */
	reviewSummary: Reviews;
	/** Reviews associated with the product. */
	reviews: ReviewConnection;
	/** Product SEO details. */
	seo: SeoDetails;
	/** Whether or not the cart call to action should be visible for this product. */
	showCartAction: Scalars['Boolean']['output'];
	/** Default product variant when no options are selected. */
	sku: Scalars['String']['output'];
	/** Type of product, ex: physical, digital */
	type: Scalars['String']['output'];
	/** Universal product code. */
	upc?: Maybe<Scalars['String']['output']>;
	/** Variants associated with the product. */
	variants: VariantConnection;
	/** Warranty information of the product. */
	warranty: Scalars['String']['output'];
	/** Weight of the product. */
	weight?: Maybe<Measurement>;
	/** Width of the product. */
	width?: Maybe<Measurement>;
};
// ...
export type ProductAttributeSearchFilterInput = {
	/** Product attributes */
	attribute: Scalars['String']['input'];
	/** Product attribute values */
	values: Array<Scalars['String']['input']>;
};
// ...
export type ProductConnection = {
	__typename?: 'ProductConnection';
	/** Collection info */
	collectionInfo?: Maybe<CollectionInfo>;
	/** A list of edges. */
	edges?: Maybe<Array<Maybe<ProductEdge>>>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/** An edge in a connection. */
export type ProductEdge = {
	__typename?: 'ProductEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String']['output'];
	/** The item at the end of the edge. */
	node: Product;
};
// ...
export type RatingSearchFilterInput = {
	/** Maximum rating of the product. */
	maxRating?: InputMaybe<Scalars['Float']['input']>;
	/** Minimum rating of the product. */
	minRating?: InputMaybe<Scalars['Float']['input']>;
};
// ...
export type SearchProductsFiltersInput = {
	/** Filter by products belonging to any of the specified Brands. */
	brandEntityIds?: InputMaybe<Array<Scalars['Int']['input']>>;
	/** Filter by products belonging to a single Category. This is intended for use when presenting a Category page in a PLP experience. This argument must be used in order for custom product sorts and custom product filtering settings targeted at a particular category to take effect. */
	categoryEntityId?: InputMaybe<Scalars['Int']['input']>;
	/** Filter by products belonging to any of the specified Categories. Intended for Advanced Search and Faceted Search/Product Filtering use cases, not for a page for a specific Category. */
	categoryEntityIds?: InputMaybe<Array<Scalars['Int']['input']>>;
	/** When set to True, hides products which are out of stock. Defaults to False. This filter will do nothing unless your store has the Product Filtering feature available on your plan and enabled. If it is supplied when your store does not have the feature enabled, it will be silently ignored. */
	hideOutOfStock?: InputMaybe<Scalars['Boolean']['input']>;
	/** Filters by Products which have explicitly been marked as Featured within the catalog. If not supplied, the Featured status of products will not be considered when returning the list of products. */
	isFeatured?: InputMaybe<Scalars['Boolean']['input']>;
	/** Filters by Products which have explicit Free Shipping configured within the catalog. If not supplied, the Free Shipping status of products will not be considered when returning the list of products. */
	isFreeShipping?: InputMaybe<Scalars['Boolean']['input']>;
	/** Search by price range. At least a minPrice or maxPrice must be supplied. */
	price?: InputMaybe<PriceSearchFilterInput>;
	/** Filter by the attributes of products such as Product Options and Product Custom Fields. This filter will do nothing unless your store has the Product Filtering feature available on your plan and enabled. If it is supplied when your store does not have the feature enabled, it will be silently ignored. */
	productAttributes?: InputMaybe<Array<ProductAttributeSearchFilterInput>>;
	/** Filter by rating. At least a minRating or maxRating must be supplied. This filter will do nothing unless your store has the Product Filtering feature available on your plan and enabled. If it is supplied when your store does not have the feature enabled, it will be silently ignored. */
	rating?: InputMaybe<RatingSearchFilterInput>;
	/** Boolean argument to determine whether products within sub-Categories will be returned when filtering products by Category. Defaults to False if not supplied. */
	searchSubCategories?: InputMaybe<Scalars['Boolean']['input']>;
	/** Textual search term. Used to search for products based on text entered by a shopper, typically in a search box. Searches against several fields on the product including Name, SKU, and Description. */
	searchTerm?: InputMaybe<Scalars['String']['input']>;
};

/** Sort to use for the product results. Relevance is the default for textual search terms, and “Featured” is the default for category page contexts without a search term. */
export enum SearchProductsSortInput {
	AToZ = 'A_TO_Z',
	BestReviewed = 'BEST_REVIEWED',
	BestSelling = 'BEST_SELLING',
	Featured = 'FEATURED',
	HighestPrice = 'HIGHEST_PRICE',
	LowestPrice = 'LOWEST_PRICE',
	Newest = 'NEWEST',
	Relevance = 'RELEVANCE',
	ZToA = 'Z_TO_A',
}

/** The Search queries. */
export type SearchQueries = {
	__typename?: 'SearchQueries';
	/** Details of the products and facets matching given search criteria. */
	searchProducts: SearchProducts;
};

/** The Search queries. */
export type SearchQueriesSearchProductsArgs = {
	filters: SearchProductsFiltersInput;
	sort?: InputMaybe<SearchProductsSortInput>;
};